requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
dj-database-url==2.2.0
dj-rest-auth==6.0.0
Django==5.0.4
django-heroku==0.3.1
django-allauth==0.62.1
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
idna==3.7
mysqlclient==2.2.4
packaging==24.1
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
